Dart Board Assignment
Assignment: Your task in this project is to construct and perform a Monte Carlo simulation of a darts game. Make an assumption about the distribution of how darts hit on the board. Write an algorithm, and code it in MATLAB. Run 1000 simulations to determine the mean score for throwing 5 darts.